+++ src/doomstat.c
@@ -164,6 +164,7 @@ int no_radsuit_tint;
 int damagecount_cap;
 int bonuscount_cap;
 int fake_contrast;
+int diminished_lighting; // CFG-Only
 int wipe_speed_percentage;
 int s_clipping_dist_x2;
 int one_key_saveload;