+++ src/doomstat.h
@@ -492,6 +492,7 @@ extern int no_radsuit_tint;
 extern int damagecount_cap;
 extern int bonuscount_cap;
 extern int fake_contrast;
+extern int diminished_lighting; // CFG-Only
 extern int wipe_speed_percentage;
 extern int s_clipping_dist_x2;
 extern int one_key_saveload;